requirements.txt
openai>=1.0.0
anthropic>=0.28.0
google-generativeai>=0.3.0
requests>=2.31.0
yfinance>=0.2.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
argparse>=1.4.0
datetime>=5.0
pandas>=2.0.0
numpy>=1.24.0
feedparser>=6.0.0